Reimplement parseJSON using recursive descent parser
Added functions to properly handle objects and arrays and fixed
the bugs of all the other functions previously added (but were not
yet used by parseJSON).

The first prototype of parseJSON was a hackish version that did not
properly follow JSON.parse rules but managed to pass all test cases.
This meant that the program would not be able to identify invalid JSON
strings and handle special cases. As a result, it was necessary to
reimplement the function.

Now, the function follows the structure as shown on json.org, properly
parsing objects, arrays, values, strings, and numbers. However, due
to the implementation, the index that the error messages specify
are not yet correct.